feat: added netaddr

+++ Dockerfile
@@ -5,8 +5,8 @@ VOLUME /etc/ansible
 VOLUME /usr/share/ansible
 
 RUN apk add --update-cache --quiet git ansible docker-py openssh-client \
-  py3-dnspython ansible-lint helm py3-kubernetes py3-jsonpatch py3-yaml && \
-  rm -rf /var/lib/apk/db /var/cache/apk/*
+  py3-dnspython ansible-lint helm py3-kubernetes py3-jsonpatch py3-yaml \
+  py3-netaddr && rm -rf /var/lib/apk/db /var/cache/apk/*
 
 # Copy in our entrypoint scripts
 COPY scripts /
